approval_engine/requirement.txt
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
cryptography==41.0.5
DateTime==5.2
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
furl==2.1.3
idna==3.4
logging==0.4.9.6
numpy==1.25.2
opencv-python==4.8.0.74
orderedmultidict==1.0.1
oscrypto==1.3.0
packaging==23.2
Pillow==10.0.0
platformdirs==3.9.1
psycopg2==2.9.6
pycparser==2.21
pycryptodomex==3.19.0
PyJWT==1.7.1
pyOpenSSL==23.2.0
pytz==2023.3
requests==2.28.1
secure-smtplib==0.1.1
six==1.16.0
snowflake-connector-python==3.3.1
snowflake-ingest==1.0.5
sortedcontainers==2.4.0
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
virtualenv==20.24.2
zope.interface==6.1
